payments_analysis/requirement.txt
pymongo>=3.1
python-dateutil==2.2
datetime
